climate_data_pipeline/requirements.txt
pandas
requests
google-cloud-storage
dlt
apache-airflow-providers-google
